+++ Installer.h
@@ -1,8 +1,6 @@
 #pragma once
 #include "pch.h"
 
-#include "Helpers.h"
-
 namespace NppShell::Installer
 {
     HRESULT RegisterOldContextMenu();
