+++ NppShell.vcxproj
@@ -258,7 +258,7 @@
     <ClInclude Include="EditWithNppExplorerCommandHandler.h" />
     <ClInclude Include="ExplorerCommandBase.h" />
     <ClInclude Include="framework.h" />
-    <ClInclude Include="Helpers.h" />
+    <ClInclude Include="PathHelper.h" />
     <ClInclude Include="Installer.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="SimpleFactory.h" />
@@ -267,7 +267,7 @@
     <ClCompile Include="dllmain.cpp" />
     <ClCompile Include="EditWithNppExplorerCommandHandler.cpp" />
     <ClCompile Include="ExplorerCommandBase.cpp" />
-    <ClCompile Include="Helpers.cpp" />
+    <ClCompile Include="PathHelper.cpp" />
     <ClCompile Include="Installer.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
