+++ PathHelper.h
